修改刷新时间间隔，防止每次切换频道一直刷新

+++ TTNewsSDK/TTNewsSDK/Classes/Expand/Tools/TTTool.h
@@ -9,6 +9,6 @@
 #import <Foundation/Foundation.h>
 
 @interface TTTool : NSObject
-+ (BOOL )lastUpdateTimeWithChannelName:(NSString *)channelName withinHours:(NSInteger )hours;
++ (BOOL )lastUpdateTimeWithChannelName:(NSString *)channelName withinHours:(CGFloat )hours;
 + (void)updateFetchTime:(NSString *)channelName;
 @end
